backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File Handling & Storage
python-magic==0.4.27
boto3==1.29.3
Pillow==10.1.0

# Utilities
pydantic==2.5.1
email-validator==2.1.0.post1
python-dotenv==1.0.0
requests==2.31.0

# Caching & Performance
redis==5.0.1
celery==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Internationalization
babel==2.13.1
pytz==2023.3.post1

# API Documentation
swagger-ui-bundle==0.0.9

# Development Tools
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Monitoring & Logging
sentry-sdk==1.32.0
prometheus-client==0.18.0